requirements.txt
django
djangorestframework
django-cors-headers
django-allauth
psycopg2-binary
requests
jwt
PyJWT --upgrade